+++ src/io_util.cpp
@@ -475,10 +475,10 @@ void ssrlcv::writeCSV(float* values, int num, std::string filename){
 
 /*
  * Takes in a c++ vector and prints it all on one line of a csv
- * @param v a vector of unsinged long long int guys
+ * @param v a vector of float guys
  * @param filename a string representing the desired filename of the csv output
  */
-void ssrlcv::writeCSV(std::vector<unsigned long long int> v, std::string filename){
+void ssrlcv::writeCSV(std::vector<float> v, std::string filename){
   std::ofstream outfile;
   outfile.open("out/" + filename + ".csv");
   for (int i = 0; i < v.size(); i++) outfile << v[i] << ",";
